refactor: move WebWrapper files from @calcom/atoms to @calcom/web
This is the first step in breaking the circular dependency between
@calcom/atoms and @calcom/web. The goal is to ensure @calcom/atoms
has no knowledge of @calcom/web, while @calcom/web can import from
@calcom/atoms.

Changes:
- Move EventTypeWebWrapper and related WebWrapper files from
  @calcom/atoms to @calcom/web/modules/event-types/wrappers/
- Update imports in @calcom/web to use new wrapper locations
- Modify EventSetupTab to accept isPlatform as prop instead of hook
- Modify Locations component to accept isPlatform as prop
- Remove EventTypeWebWrapper export from @calcom/atoms package.json
