+++ apps/web/modules/event-types/wrappers/EventLimitsTabWebWrapper.tsx
@@ -1,0 +1,8 @@
+import { EventLimitsTab } from "../components/tabs/limits/EventLimitsTab";
+import type { EventLimitsTabProps } from "../components/tabs/limits/EventLimitsTab";
+
+const EventLimitsTabWebWrapper = (props: EventLimitsTabProps) => {
+  return <EventLimitsTab {...props} />;
+};
+
+export default EventLimitsTabWebWrapper;
