+++ apps/web/modules/event-types/wrappers/EventRecurringWebWrapper.tsx
@@ -1,0 +1,8 @@
+import type { EventRecurringTabProps } from "../components/tabs/recurring/EventRecurringTab";
+import { EventRecurringTab } from "../components/tabs/recurring/EventRecurringTab";
+
+const EventRecurringWebWrapper = (props: EventRecurringTabProps) => {
+  return <EventRecurringTab {...props} />;
+};
+
+export default EventRecurringWebWrapper;
