+++ apps/web/modules/event-types/wrappers/index.ts
@@ -1,0 +1,12 @@
+export { default as EventSetupTabWebWrapper } from "./EventSetupTabWebWrapper";
+export { default as EventAdvancedWebWrapper } from "./EventAdvancedWebWrapper";
+export { default as EventLimitsTabWebWrapper } from "./EventLimitsTabWebWrapper";
+export { default as EventRecurringWebWrapper } from "./EventRecurringWebWrapper";
+export { default as EventAvailabilityTabWebWrapper } from "./EventAvailabilityTabWebWrapper";
+export { default as EventTeamAssignmentTabWebWrapper } from "./EventTeamAssignmentTabWebWrapper";
+export { EventTypeWebWrapper } from "./EventTypeWebWrapper";
+export type { EventTypeWebWrapperProps } from "./EventTypeWebWrapper";
+export type {
+  EventAvailabilityTabWebWrapperProps,
+  GetAllSchedulesByUserIdQueryType,
+} from "./EventAvailabilityTabWebWrapper";
