--- packages/platform/atoms/event-types/wrappers/EventLimitsTabWebWrapper.tsx
@@ -1,8 +1,0 @@
-import { EventLimitsTab } from "@calcom/web/modules/event-types/components/tabs/limits/EventLimitsTab";
-import type { EventLimitsTabProps } from "@calcom/web/modules/event-types/components/tabs/limits/EventLimitsTab";
-
-const EventLimitsTabWebWrapper = (props: EventLimitsTabProps) => {
-  return <EventLimitsTab {...props} />;
-};
-
-export default EventLimitsTabWebWrapper;
